Move all client DB keys into 0x2 namespace
Had to make ln client's db module public to do this. We should make
it private again once the user and gateway clients are unified.
